fix: minor bugs
