+++ android/app/build.gradle
@@ -100,7 +100,7 @@ android {
         applicationId "com.WahegurooNetwork.SundarGutka"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
-        versionCode 141
+        versionCode 143
         versionName "5.8.1"
     }
     splits {
