+++ android/app/release/output-metadata.json
@@ -11,7 +11,7 @@
       "type": "SINGLE",
       "filters": [],
       "attributes": [],
-      "versionCode": 141,
+      "versionCode": 142,
       "versionName": "5.8.1",
       "outputFile": "app-release.apk"
     }
