+++ src/AboutScreen/hooks/useHeader.js
@@ -2,6 +2,7 @@ import React, { useEffect } from "react";
 import { Icon } from "@rneui/themed";
 import colors from "../../common/colors";
 import { styles } from "../styles";
+import STRINGS from "../../common/localization";
 
 const useHeader = (navigation) => {
   const { WHITE_COLOR } = colors;
@@ -12,6 +13,7 @@ const useHeader = (navigation) => {
   );
   useEffect(() => {
     navigation.setOptions({
+      title: STRINGS.about,
       headerTitleStyle,
       headerStyle,
       headerLeft,
