+++ src/EditBaniOrder/styles.js
@@ -15,9 +15,6 @@ export const styles = StyleSheet.create({
     fontSize: 24,
     textAlign: "center",
   },
-  gestureBackColor: {
-    backgroundColor: colors.LABEL_COLORS,
-  },
 });
 
 export const activeColor = (isActive, backColor) => ({
@@ -28,5 +25,6 @@ export const nightStyles = (isNightMode) => {
   return {
     backColor: { backgroundColor: isNightMode ? colors.NIGHT_BLACK : colors.WHITE_COLOR },
     textColor: { color: isNightMode ? colors.WHITE_COLOR : colors.NIGHT_BLACK },
+    viewBackColor: { backgroundColor: isNightMode ? colors.WHITE_COLOR : colors.NIGHT_GREY_COLOR },
   };
 };
