+++ src/Settings/components/fontSize.jsx
@@ -1,7 +1,6 @@
 import React, { useState } from "react";
 import { useSelector } from "react-redux";
 import { STRINGS } from "../../common";
-import { FONT_SIZES } from "../../common/actions/constant";
 import { setFontSize } from "../../common/actions";
 import { BottomSheetComponent, ListItemComponent } from "./comon";
 
@@ -9,6 +8,13 @@ function FontSizeComponent() {
   const [isVisible, toggleVisible] = useState(false);
   const fontSize = useSelector((state) => state.fontSize);
   const fontSizeIcon = require("../../../images/fontsizeicon.png");
+  const FONT_SIZES = [
+    { key: "EXTRA_SMALL", title: STRINGS.extra_small },
+    { key: "SMALL", title: STRINGS.small_default },
+    { key: "MEDIUM", title: STRINGS.medium },
+    { key: "LARGE", title: STRINGS.large },
+    { key: "EXTRA_LARGE", title: STRINGS.extra_large },
+  ];
   return (
     <>
       <ListItemComponent
