+++ src/Settings/components/padched.jsx
@@ -2,13 +2,16 @@ import React, { useState } from "react";
 import { useSelector } from "react-redux";
 import { STRINGS } from "../../common";
 import { setPadched } from "../../common/actions";
-import { PADCHED_SETTINGS } from "../../common/actions/constant";
 import { ListItemComponent, BottomSheetComponent } from "./comon";
 
 function PadchedSettingsComponent() {
   const [isVisible, toggleVisible] = useState(false);
   const padched = useSelector((state) => state.padched);
   const padchedIcon = require("../../../images/larivaaricon.png");
+  const PADCHED_SETTINGS = [
+    { key: "SAT_SUBHAM_SAT", title: STRINGS.sat_subham_sat_default },
+    { key: "MAST_SABH_MAST", title: STRINGS.mast_sabh_mast },
+  ];
   return (
     <>
       <ListItemComponent
