+++ src/Settings/components/reminders/ReminderOptions/ReminderOptions.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useRef, useMemo } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { View, StatusBar } from "react-native";
+import { View, StatusBar, ScrollView, SafeAreaView } from "react-native";
 import Accordion from "react-native-collapsible/Accordion";
 import PropTypes from "prop-types";
 import ModalSelector from "react-native-modal-selector";
@@ -13,7 +13,7 @@ import {
   trackReminderEvent,
   updateReminders,
 } from "../../../../common";
-import { styles, accordianNightColor } from "./styles";
+import { styles, accordianNightColor, optionContainer } from "./styles";
 import { AccordianContent, AccordianHeader } from "./components";
 import { useHeader, useDefaultReminders, useFetchBani } from "./hooks";
 
@@ -29,6 +29,7 @@ function ReminderOptions({ navigation }) {
   const [baniListData, setBaniListData] = useState([]);
   const dispatch = useDispatch();
   const selector = useRef(null);
+  const { backgroundColor, color } = optionContainer(isNightMode);
 
   const accNightColor = useMemo(() => accordianNightColor(isNightMode), [isNightMode]);
   const setDefaultReminders = useDefaultReminders(setStateData);
@@ -53,6 +54,7 @@ function ReminderOptions({ navigation }) {
     if (existingObjIndex === -1) {
       const newObj = {
         key: newObjKey,
+        id: selectedOption.id,
         gurmukhi: selectedOption.gurmukhi,
         translit: selectedOption.translit,
         enabled: true,
@@ -69,38 +71,53 @@ function ReminderOptions({ navigation }) {
   };
 
   return (
-    <View
+    <SafeAreaView
       style={[
         isNightMode && { backgroundColor: colors.INACTIVE_VIEW_COLOR_NIGHT_MODE },
         styles.flexView,
       ]}
     >
-      <StatusBar backgroundColor={colors.TOOLBAR_COLOR_ALT2} barStyle="light-content" />
-      {stateData.length > 0 && (
-        <Accordion
-          activeSections={activeSections}
-          sections={stateData}
-          underlayColor={accNightColor}
-          renderHeader={(section, index, isActive) => (
-            <AccordianHeader section={section} isActive={isActive} />
+      <ScrollView>
+        <View>
+          <StatusBar backgroundColor={colors.TOOLBAR_COLOR_ALT2} barStyle="light-content" />
+          {stateData.length > 0 && (
+            <Accordion
+              activeSections={activeSections}
+              sections={stateData}
+              underlayColor={accNightColor}
+              renderHeader={(section, index, isActive) => (
+                <AccordianHeader section={section} isActive={isActive} />
+              )}
+              renderContent={(section, index, isActive) => (
+                <AccordianContent section={section} isActive={isActive} />
+              )}
+              onChange={updateSections}
+            />
           )}
-          renderContent={(section, index, isActive) => (
-            <AccordianContent section={section} isActive={isActive} />
-          )}
-          onChange={updateSections}
-        />
-      )}
-      <ModalSelector
-        data={reminderBaniData}
-        cancelText={STRINGS.cancel}
-        optionTextStyle={styles.modalSelecText}
-        onChange={(option) => {
-          createReminder(option);
-        }}
-        customSelector={<View />}
-        ref={selector}
-      />
-    </View>
+          <ModalSelector
+            supportedOrientations={[
+              "portrait",
+              "portrait-upside-down",
+              "landscape-left",
+              "landscape-right",
+            ]}
+            // style={{ backgroundColor: "#000" }}
+
+            data={reminderBaniData}
+            cancelText={STRINGS.cancel}
+            optionTextStyle={{ ...styles.modalSelectText, color }}
+            onChange={(option) => {
+              createReminder(option);
+            }}
+            customSelector={<View />}
+            ref={selector}
+            cancelTextStyle={{ color }}
+            cancelStyle={{ backgroundColor }}
+            optionContainerStyle={{ backgroundColor }}
+          />
+        </View>
+      </ScrollView>
+    </SafeAreaView>
   );
 }
