+++ src/common/localization.js
@@ -128,9 +128,9 @@ const STRINGS = new LocalizedStrings({
     about: "हमारे बारे",
     ABOUT_WELCOME: "हम आप के सुझाव और टिप्पणिओं का स्वागत करते हैं।",
     ABOUT_HELP: "और जानकारी, मदद या सुझाव के लिए यहां जाएँ।",
-    ABOUT_RESPECT: "एप्प का इस्तेमाल करने से पहले सत्कार सहत अपने जूते उतार के सिर ढक लें",
-    ABOUT_SG: "सुंदर गुटका इस्तेमाल करता है",
-    ABOUT_OPEN_SOURCE: "-ओपन सोर्स डाटाबेस और बहुत सारी गुरबाणी एप्प जैसे SikhiToTheMax",
+    ABOUT_RESPECT: "एप्प का इस्तेमाल करने से पहले सत्कार सहत अपने जूते उतार के सिर ढक लें।",
+    ABOUT_SG: "सुंदर गुटका इस्तेमाल करता है।",
+    ABOUT_OPEN_SOURCE: "-ओपन सोर्स डाटाबेस और बहुत सारी गुरबाणी एप्प जैसे SikhiToTheMax।",
     ABOUT_PARDON: "भूलों की क्षमा !",
     above_raag_headings: "राग शीर्षक",
     anmol_lipi: "अनमोल लिपि",
@@ -250,10 +250,10 @@ const STRINGS = new LocalizedStrings({
     about: "ਸਾਡੇ ਬਾਰੇ",
     ABOUT_WELCOME: "ਅਸੀਂ ਤੁਹਾਡੀਆਂ ਟਿੱਪਣੀਆਂ ਅਤੇ ਸੁਝਾਵਾਂ ਨੂੰ ਖਿੜੇ ਮੱਥੇ ਪ੍ਰਵਾਨ ਕਰਦੇ ਆ।",
     ABOUT_HELP: "ਹੋਰ ਜਾਣਕਾਰੀ, ਸੁਝਾਅ ਯਾ ਮੱਦਦ ਲਈ ਏਥੇ ਜਾਓ",
-    ABOUT_RESPECT: "ਐਪ ਨੂੰ ਵਰਤਣ ਲੱਗੇ ਸਤਿਕਾਰ ਸਹਿਤ ਆਪਣੇ ਜੋੜੇ ਲਾਹ ਕੇ ਸਿਰ ਢੱਕ ਲਓ",
+    ABOUT_RESPECT: "ਐਪ ਨੂੰ ਵਰਤਣ ਲੱਗੇ ਸਤਿਕਾਰ ਸਹਿਤ ਆਪਣੇ ਜੋੜੇ ਲਾਹ ਕੇ ਸਿਰ ਢੱਕ ਲਓ।",
     ABOUT_SG: "ਸੁੰਦਰ ਗੁਟਕਾ ਵਰਤਦਾ ਹੈ",
     ABOUT_OPEN_SOURCE:
-      "- ਓਪਨ ਸੋਰਸ ਗੁਰਬਾਣੀ ਡਾਟਾਬੇਸ ਅਤੇ ਬਹੁਤ ਸਾਰੀਆਂ ਗੁਰਬਾਣੀ ਐਪਾਂ ਜਿਵੇਂ ਕੇ SikhiToTheMax",
+      "- ਓਪਨ ਸੋਰਸ ਗੁਰਬਾਣੀ ਡਾਟਾਬੇਸ ਅਤੇ ਬਹੁਤ ਸਾਰੀਆਂ ਗੁਰਬਾਣੀ ਐਪਾਂ ਜਿਵੇਂ ਕੇ SikhiToTheMax।",
     ABOUT_PARDON: "ਭੁੱਲ ਚੁੱਕ ਮਾਫ !",
     above_raag_headings: "ਰਾਗ ਸਿਰਨਾਵੇਂ",
     anmol_lipi: "ਅਨਮੋਲ ਲਿਪੀ",
@@ -373,14 +373,14 @@ const STRINGS = new LocalizedStrings({
     ABOUT_WELCOME: "Veuillez envoyer vos commentaires, suggestions et corrections!",
     ABOUT_HELP: "Pour des informations, des suggestions ou de l'aide, visitez-nous à",
     ABOUT_RESPECT:
-      "Veuillez vous couvrir la tête et retirer vos chaussures par respect lorsque vous utilisez cette application",
+      "Veuillez vous couvrir la tête et retirer vos chaussures par respect lorsque vous utilisez cette application.",
     ABOUT_SG: "Sundar Gutka utilise",
     ABOUT_OPEN_SOURCE:
       "- la base de données open source gurbani et l'api utilisées dans de nombreuses applications gurbani, telles que SikhiToTheMax.",
     ABOUT_PARDON: "Pardonnez les erreurs [Boule Chouc Màf]",
     above_raag_headings: "Au-Dessus des rubriques raag",
     anmol_lipi: "Anmol Leppi",
-    APP_VERSION: "Version de l'Application",
+    APP_VERSION: "Version",
     AUTO_SCROLL: "Défilement Automatique",
     bani_length: "Longueur de Bani",
     bani_length_alert_1:
@@ -506,7 +506,7 @@ const STRINGS = new LocalizedStrings({
     ABOUT_PARDON: "Perdona gli errori [Bùll Ciùc Màf]",
     above_raag_headings: "Sopra i titoli raag ",
     anmol_lipi: "Anmòl Leppi",
-    APP_VERSION: "Versione dell'App",
+    APP_VERSION: "Versione",
     AUTO_SCROLL: "Scorrimento Automatico",
     bani_length: "Lunghezza delle Bani",
     bani_length_alert_1:
@@ -630,7 +630,7 @@ const STRINGS = new LocalizedStrings({
     ABOUT_PARDON: "Que se perdone los errores [Búl Chúk Máf]",
     above_raag_headings: "Por encima de los títulos raag ",
     anmol_lipi: "Anmól Lippi ",
-    APP_VERSION: "Versión de la aplicación",
+    APP_VERSION: "Versión ",
     AUTO_SCROLL: "Desplazamiento automático",
     bani_length: "Largueza de las Bani",
     bani_length_alert_1:
